week3 step2 인스턴스 소통 구현, 추가 디벨롭 예정
